assists: baremetal_gentestapp: Check the interrupt-parent of IPs before adding their intr example in peripheral test
When the driver example depends on the interrupt property of the IP, check the
validity of the interrupt-parent in the node before pulling it in the peripheral
test. There are use cases with PS+PL designs where the PS IPs are mapped to the
MB but the ps_pl interrupt connections are missing. In such cases, interrupt
dependent examples should not be pulled in the peripheral tests.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>