tests: add selftest for assist sanity checking
Along with the built-in tests of lopper_sanity, having
reference self tests for assists is useful for testing
extended functionality that happens outside of lop files.

Much of the contents of lopper_sanity.py will migrate
to this newly created selftest directory.

For now, it is only an external assist sanity check that
is exercising node movement and manipulation.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>