lopper: search configured paths for assists
Previously lopper would only search for assists in directories
relative to the main lopper executable/script.

This is not condusive to an installed lopper that will not
have writable subdirectories (not should it). Add the ability
too add to the assist search path through directories that
are in the main SDT class.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>