lopper/tree: fix node addition and move robustness
While is was previously possible to move a node that was
already part of a tree, care needed to be take to delete
the old node and then add the new one.

To allow more use of this functionality in assists, we
add checks and adjust tree operations to detect moves and
perform delete/add automatically.

Additionally, more operations take into account parent
and child nodes to allow entire subtrees to be moved
and manipulated.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>