tree: fix property debug level and output
The debug level increase of properties after deepcopy
was left by mistake.

But that debug level increase exposed a debug output that
has been incorrect for some time, and causes a stack trace
when scalar inputs are read and put into the tree (i.e. yaml).

We drop the debug increase, and also fix the debug statement
so it won't cause issues in the future.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>