tree: add API to reorder nodes
Nodes are kept in their read order to be compatible with the
way that libfdt reads and manages dtb.

The side effect of this means that if a new node is added, it
is almost always added last. While this doesn't matter when
processing the device tree programatically, it does make a
difference when reading/printing device trees.

To allow finer grained ordering of child nodes, we add a new
API: reorder_child()

This allows child nodes of a given parent to be shuffled and
ordered by an assist.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>