modify: delete node before copy and rename
commit 41d9d99f38bf49ff72ee7257ee3a77faccd7727b changed tree
ooperations to detect phandle collisions and perform fixups
to avoid duplication.

The modify lop is not hitting this condition since it is only
deleting the old node AFTER the new one is added. It was
always a better idea to delete, copy, add a moved node, but
it previously wasn't an issue.

By deleting the node first, we avoid the phandle collision and
old nodes with phandle references to the old node continue
to be valid.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>