tree: fix phandle collision detection
The ability to detect and fix phandle collisions was added
to ensure tree sanity when an assist was used to copy a
node, update and then re-add it to the tree. The check was
added so that the user would not have to consider the global
phandle space.

BUT

The node.load() routine was assigning the phandle in a manner
that was also adding it to the tree. As such, all copied nodes
were being detected as collisions even if the old node was
deleted from the tree.

We move the collsion check before node.load() and we'll no
longer have new phandles generated that in turn break references
throughout the tree.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>