lopper: update default phandle map definitions
The definition for interrupt-map was not in compliance
with the ePapr definition of the property. We adjust it
to match the lookups that are required.

At the same time, we can add some new definitions and
adjust definitions that needed post processing (+1) since
phandles are now accounted for in field processing and
we no longer need to adjust by default.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>