assists: gen_domain_dts: Add ocm IPs into linux ignore list
With taxonomy changes, a generic ocm_ram name was added in the get_memranges API.
This pattern is matching with OCM IPs of zynqmp and Versal and leading to the
presence of ocm node in linux device trees which is not desired. Add all the
OCM IPs in the linux ignore list to avoid the same.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>