lopper: fix property comparisions
The reworked phandle map was not returning "#invalid" for
phandles that could not be located in the tree.

This broke an assumption in property compares (and hence
lop files), since they are looking for that string to
determine a phandle. Since property compares are done on
newly created / temp properties, no phandles can be
resolved as the property isn't in a tree. This meant that
selection was broken based on property.

We use "None" instead of zero to signify a failed lookup
and we can keep all use cases as they were.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>