lopper: add strict mode to nodes()
nodes() is the exception safe way to search for nodes
in a tree. But unlike a direct path lookup, nodes() can
return more than one match.

To avoid this add an optional "strict" argument to the
routine which will do a bounded regex match, ensuring that
we will find a more specific match.

We can also use this in the modify lop, as a lookup is
done when doing phandle replacements.

To get specific / strict matching in the modify lop, the
newly introduced "flags" property can be used with "strict"
in a space seperated list in the string.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>