lopper: flush, unify and simplify output (print())
The LopperTreePrinter() class was the initial implementation
of dts writing via lopper code. It used the callback mechanism
of the base LopperTree to coordinate the various elements of
a print.

This implementation made it difficult to print a single node
or a property. To allow that, print() methods were added to
each object. Those print methods were identical to the callbacks.

We've had to keep the techniques in sync since that new way
of printing was introduced.

If we drop the callbacks from LopperTreePrinter, we can then
use the print() routines from the objects and end with the
same result.

Along with this code removal, it was noted that some output
is NOT flushed to the output stream in a timely manner.
Adding 'flush' to the print() routines solves that issue.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>