lopper: add heuristics to string format detection
It was found that device trees with properties of the format:

    mem-ctrl-base-address =  <0x76000000>;

would be detected as strings, which results in obviously invalid
output.

Even when we add a scheme based type validation, we still have
an issue with dts or dtb files that do not have an associated
schema.

We add two more checks to string detection to ensure that we aren't
miscategorizing numeric values.

  - if the length is 4 or 8, we try a libfdt decode into a number
    as a final check. If that passes, we declare that it isn't a
    string and let numeric processing proceed.

This also adds a lopper sanity check to test for this condition
in the future.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>