lopper/device tree: variable sized records
When processing the cells of a property, lopper was initially
optimized to assume that all record sizes were the same. This
allowed a single pass through the definition to calculate the
size of each record, and then simply chunking up the property
for processing.

This breaks properties that do lookups of values to determine
their size. Which can return different sized groupings of
values to form the records of a property.

With this change, we now walk the entire property doing lookups
to build a list of group sizes. Those group sizes are used
to chunk the property for processing (in particular phandle
replacement).

A sanity check to the "chunking" operation has also been added
to ensure that the sum of the group sizes equals the length
of the property. This ensures that bad lookups or defaults
don't incorrectly resolve to phandles.

We also add a few more definitions to our phandle map, to allow
them to be symbolically replaced and properly test this
fixed functionality.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>