lopper: make output file type detection more specific
The output file "type" handling of lopper was written quite some
time ago, and didn't have fully specified and anchored regex's.

As a result, paths with some of the extensions embedded in them
could be mistaken for an output file type.

We tighten up the regex's and make the output directory handling
consistent across all types.

A more standard extension handling mechanism can be introduced
in the future.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>