assists:baremetallinker Add pmc_ram in valid list of memory IPs
Add pmc_ram in valid list of memory IPs. This is needed to support the static
SDT specifically targeted to build SpartanUp PLM. It will have only pmc_ram memory
IP in it.

This wont cause issues for other platforms because we are not generating pmc_ram
as memory IP today and these are hardcoded in the linker section.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>