lops/assists: Rename APU gic labels to gic instead of gic_a*
Linux device trees traditionally has only one gic reference in them. System
Device Tree has imux to denote the presence of both APU (say gic_a72) and RPU
(say gic_r5) gics. This imux is currently replaced with the APU gic node label
(gic_a72) in linux device tree with the help of lops and assists. In case when
the user already has an old existing dt overlay for linux domain, which involves
gic as an interrupt-parent, the expected node reference for the interrupt-parent
would be gic instead of gic_a72 as per the traditional linux device tree
assumptions.

Update the lops and assists files to replace imux with gic instead of gic_a72
kind of labels to make the nodes uniform in the traditional linux device trees
and the linux device trees generated with system device tree based flow.
This will help in smooth application of the old linux overlay files.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>