baremetallinker: set memtest_config to False while updating default_ddr based on valid_mem_ips
Set memtest_config to False when default_ddr is updated based on valid_memips
so that the rest of the logic doesnt evaluate default_ddr again for memory test
configuration.

Also set a flag valid_memfix to inform rest of the logic to not set ddr_default
further as the default mem was already set using valid mem ips.
Use the same valid_memfix flag to pop the default_ddr to zeroth location in memory
IP lists so that the memory test doesnt list it in memories to be tested.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Acked-by: Anirudha Sarangi <anirudha.sarangi@amd.com>