lopper: add search paths for input files
assists could be found in non-standard or user defined locations
via the -A option. No such method exists for input files (which
can be lop files, or device trees).

This commit adds the -I parameter to lopper, which functions
exactly the same as -A (and in fact IS the same option
internall).

We also introduce and environment variable: LOPPER_INPUT_DIRS
That can be used to specify search directories when the command
line options are not convienient.

There are more internal changes than simply adding the new option
as the asisst and input file locating are unified, as are the
types of input checking.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>