lopper: baremetalconfig_xlnx: Add support for 'mdioproducer-baseaddr' property
This patch adds support for generating the `mdioproducer-baseaddr`

if the `phy-handle` property is absent, default to `0x0` as the base address.

if PHY handle under MAC node,if the parent node has a `reg` property,
it is considered a MAC node. Extract the base address from its register.

if PHY handle under MDIO node,if the parent node lacks a `reg` property but
its grandparent has one, treat it as an MDIO node and extract the base address
from the grandparent.

Signed-off-by: Venkatesh Odela <venkatesh.odela@amd.com>