tree: ensure that properties with a value have a type
LopperProperties typically get a type when the property
is resolved against a tree.

but! it is possible to create a property and then use
it in a test before it has ever been resolved. In that
case, the lack of a type causes issues and can cause
a false positive or negative.

When a property is created with a value, ensure that
a type is guessed for it, such that we know it will
always be defined.

We also put a log in the selection/comparision code
so that we can clearly see any future occurences of
undefined/invalid types.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>