assists/domain_access: enable command line asisst mode
Allow domain_access to be called as part of a lopper commandline
in pipeline mode. To do this, we add parameter handling so that
the target node can be passed.

The following is an example of the assist being run as an auto
load and run assist:

 % lopper.py -f --enhanced -t "/domains/Secure" -a domain_access --auto -v  -x '*.yaml' -i domains.yaml sdt.dts test.dts

And as a command line pipeline:

 % lopper.py -f --enhanced -x '*.yaml' -i domains.yaml sdt.dts test.dts -- domain_access -v -v -t /domains/Secure

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>