tree: add dts_overlay type
Add a _type attribute to LopperTrees. This allows us to
output a different preamble/indentation framework based
on the type of the tree.

Everything by default is of type "dts", but "dts_overlay"
can be assigned to the _type to enable special processing.

Note: the output extension code does not consult _type
(yet), so there's no need for "json" and "yaml" tree types
as the output filename extension distinguish them.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>