lopper/tree: make dts input node merging configurable
Allow the user to specify what to do with repeated nodes in SDT
support files (of any format).

The cfg parser setting: --cfgval dts.merge=all

Is now checked when adding a node from a support file. If merging
is enabled, the core tree routines will be instructed to merge
nodes.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>