lopper: allow multiple independenty setup/print calls
When generating a device tree overlay file, we may need to
run lopper operations on the SDT multiple times. To ensure
this is possible, we reset the lops/assists and other
registered handlers before looking at new ones in setup.

We also stop using the dedicated LopperTreePrinter() as
we must ensure that properties are copied to it for
correct output (i.e. tree type). This can be error prone,
so we just use the existing tree and call print().

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>