lops: lop-microblaze-riscv: Add support for b extension in arch flags
Add a "b" flag to the architecture flags when the ZBA, ZBB, and ZBS
extensions are enabled in the hardware design. If any of these
extensions are not enabled, continue using the legacy _zba, _zbb, and
_zbs flags.

For now, continue using the legacy libraries to avoid the library mapping
logic, as there are only a limited number of "b" extension-enabled libraries
available in the Vitis installation.

Signed-off-by: Mubin Sayyed <mubin.sayyed@amd.com>