lopper:__main__.py: Fixed file not exist in the binary flow
PyInstaller extracts the bundled application to a temporary directory like /tmp/_MEIXXXX/.
When importing a package (_init_.py file), Python treats it as a module and assigns it a
path inside /tmp/_MEIXXXX/lopper/ (because it keeps packages in their directories).
When running _main_.py, its path is just /tmp/_MEIXXXX/ because it’s the main entry point.
Add import lopper_directory from the lopper

Signed-off-by: Sathish Kumar Kamishettigari <sathishkumar.kamishettigari@amd.com>