lopper:assists:baremetallinker_xlnx: Remove reset vector check for microblaze_riscv
The linker script is currently selecting the BRAM, even though DDR is available.
At this moment, the linker is validating the reset vector values.
If the reset vector values match the starting memory address values, it usually returns the DDR address.
To resolve this issue, remove the reset vector check for the microblaze_riscv,
as this is already managed in the existing code.

Signed-off-by: Sathish Kumar Kamishettigari <sathishkumar.kamishettigari@amd.com>