lopper: improve tmpfile handling
Rather than assuming that either the working directory, or the
input file directories are writeable, we can create a temporary
directory and use it for intermediate files.

Passing --save-temps will ensure that those files remain for
debugging.

While implementing this feature, it was found that some tmpfile
handling was incorrect and / or tmp files were being left uncleaned.
We improve the cleanup routines to handle these new and existing
temporary files.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>