assists: gen_domain_dts: Delete the afi (xlnx, afi-fgpa) and clocking (xlnx, fclk) nodes from linux dt
94fd51e1ca2b (assists: gen_domain_dts: Keep the nodes having no base address in
linux dt) added logic to keep all the nodes in linux dt that do not have base
addresses. Some of the nodes like afi/clocking which are part of SDT and fall
under the same category are expected to be deleted in the linux dt. Therefore,
add bruteforce logic to delete these two nodes from linux dt.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>