lopper: allow priority on embedded lops
When embedded lops are found the priority field won't be
present since the root node of the embedded lops is merged
with the root node of the tree (and we don't want to have
our lop information there.

To allow a priority on an embedded lop, check the /lops
node for a priority and promote it to the root node of
the tree we create for them.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>