sanity: add embedded lop and phandle replacment sanity test (example)
Add a test that shows how a dts file can carry instructions to lopper
on how to intepret the properties found in the dts file itself.

In this example we show that on the first run the phandle is left as
a number and on the second it is converted to a symbolic reference.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>