lopper:assists:baremetal_getsupported_comp_xlnx: Add support to include the examples in the lib_list.yaml
In the current implementation the examples are not included in the lib_list yaml file,
Add support to include the examples in the lib_list.yaml to show the available examles for the library in the GUI.
To enhance functionality, add support for reading examples from custom configurations in YAML files.
If a YAML file contains a condition within the examples section in the form of a Python code block,
read and execute this code to update the examples based on the specified conditions.
This approach allows dynamic configuration adjustments based on predefined logic,
improving flexibility and customization in managing examples

Signed-off-by: Sathish Kumar Kamishettigari <sathishkumar.kamishettigari@amd.com>