lopper:assists: Add support to read the custom configurations in the YAML
Add support for reading depends from custom configurations in YAML files.
If a YAML file contains a condition within the depends section in the form of a Python code block,
read and execute this code to update the depends based on the specified conditions.
This approach allows dynamic configuration adjustments based on predefined logic,
improving flexibility and customization in managing depends

Signed-off-by: Sathish Kumar Kamishettigari <sathishkumar.kamishettigari@amd.com>