lopper: learn phandle patterns from source code
It is preferable that the base dictionary of properties describe
complex formats (such as variable sized look up values), but in
some scenarios this is not possible (and a schema is not available).

In situations like this, symbolic phandle substitutions are valuable
for downstream tools, rountrip processing or just for visual
inspection.

To allow this, we implement two different methods to track phandles
(when source files are used).

Type 1: a static phandle map

In this method the source code is analysed for &phandle references
and the node, the property and the index of each symbolic reference
are noted. The list can be later updated if the tree changes, but
that is a manual process and can still be incorrect if the property
itself is modified.

A map of this type stays with the tree by being written to an
internal node __lopper-phandles__, which gets special processing
when it is read from a dtb or a dictionary export/import.

Type 2: learned phandle patterns

At the same time the static phandle map is being created, the references
can be used to detect phandle patterns in the input. These are similar
to the base phandle descriptions that are part of lopper.

Note: this cannot pick up patterns that use variable sized fields
(i.e. #address-cells) and hence variable sized property records. But
it can effectively learn simple phandle properties.

If we want variable sized phandle learning type 1 and type 2 would
need to be combined as well as additional hard coded heuristics.

After patterns have been learned, the base phandle descriptions are
updated with NEW properties and all other processing remains the
same .. they'll go into use automatically.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>