lopper: consider arguments when filtering assist options
The way that lopper filters and peeks at command line assist
options means that additional processing is required to allow
expansion and interpretation of the options before they are
passed to an assist.

The commit which allowed an empty SDT and output device tree
to enable --help filtering to assists didn't take into account
the values passed to options.

As a result, something like -O <output dir> would have <output dir>
maintained and subsequently treated as the input SDT (which should
be the first non getopt processed command line option/value)

By consdidering both the option and the value we will pull it out
of the arguments considered for device tree and assists.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>