lopper: allow phaandle replaceent exclusions
There are some nodes that should be excluded from phandle replacment
processing.

Since we can't know all of those nodes in advance, we add a field to
the phandle map:

   "__phandle_exclude__" : [ '.*lop.*', '/__symbols__', '/aliases' ]

Like any entry in the phandle definitions, this can be programatically
or lop-file modified.

If a node is found whose absolute path matches these exclusions, then
phandle processing won't be done on it's properties.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>