lopper: make support binary checks work with no input SDT
lopper was extended to run without an input system device tree
to allow for extended help processing.

That means setup can be called without and input system device
tree and hence the support binary check runs in more scenarios
than previously.

If no input SDT has been passed, we don't need to check for
support binaries, as we won't actually be processing a tree.

While we are making the support binary check conditional, we
extend it to allow pcpp (python cpp) as well as "cpp" to meet
the preprocessor requirements (and this matches the tests that
are made just before using the preprocessor)

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>