assists: generate_config_object: Avoid generating corrupted config object
In a very rare conner case, where we don't have an IPI channel mapped
either with R5_0 or R5_1 in pcw.dtsi, the plugin was mal-functioning.
Instead of writing IPI mask of the processer, it was writing processor
name in generated config object file.
Write IPI mask instead of processer name.

Signed-off-by: Madhav Bhatt <madhav.bhatt@amd.com>