lopper: pass output directory in the lop environment
lops are run against a tree, and by design are separated from
the system device tree's configuration.

That being said, if a lop is producing output, having it follow
the passed output directory to the main lopper invokation is useful
information.

Arrange for the output directory to appear in the lop environment
as a variable "outdir"

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>