tree: add alias lookup to deref()
The node dereference routine currently checks by phandle
and then by label.

There's another option that may be used as a shorthand to
find a node instead of the path .. an alias.

If both of our first two lookups fail, check the alias and
return that if a match is found.

Note: api users can always alias_node() directly if they
      want a direct lookup and they know they are using an
      alias.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>