assists: gen_domain_dts: Add special handling for QDMA IP when it is configured as Endpoint mode
Remove unwanted QDMA properties based on YAML definitions when the node status is "okay"
and its compatible matches the driver compatibility list.

If a QDMA node is configured as a PCIe Endpoint, update its status property to
"disabled" to ensure only the required properties are present and the
Endpoint status to disabled.

Signed-off-by: Bhavana Jupalli <Bhavana.Jupalli@amd.com>