lopper/yaml: implement preemptive merging of yaml files
If multiple yaml inputs are on the command line we can preemptively
merge them into a single file rather than relying on the tree merging
that lopper core performs (we need both).

This implementation tries to use deepmerge as the first option and
then falls back to a simpler implementation that uses no libraries.

The strategy is to append lists, and merge dictionaries to allow yaml
files with the same structure to effectively be concatenated and then
be viable for loading and processing as a tree.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>