lopper: add autorun lop files
To make it easier to stage lop files that are used to process
inputs or types of inputs, we introduce a new file type of ".lop"

When --auto is passed to lopper, these files are checked and
excuted if found.

In the future these file types will replace the "xlate" mappings,
since they can implement the same functionality.

.lop files are simply lop files (which typically have a .dts extension)
that have been placed to match given inputs and be run automatically
against that input.

The format of the auto run lop is:  <pattern>.lop

To allow wider matching a simple pattern character can be in the lop
file name "%". When % is found, it means "match from this point to
the end of the input file".

So: domain%.lop would match an input file of domains.yaml but would
not match test-domain.yaml. d%.lop would again match one and not the
other

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>