lops: add autorun domain processing lop
This lop is almost identical (and will replace) the lop:

   lop-xlate-yaml.dts

Which was the previous lop for expanding domains related
yaml. The file type expansions are meant to be generic, and
hence the name of that lop is confusing. It is not for
expanding generic yaml, it is very much dedicated to expanding
domains*.yaml files.

We will eventually remove the xlate yaml file once all users
have been convered.

The only addition to the xlate lop file is a call to do
wildcard device glob expansion.

                lop_yaml_0_2 {
                      compatible = "system-device-tree-v1,lop,select-v1";
                      select_1;
                      select_2 = "/domains";
                      lop_yaml_0_2 {
                              compatible = "system-device-tree-v1,lop,code-v1";
                              inherit = "subsystem";
                              code = "
                                    for i,d in enumerate(__selected__):
                                        subsystem.wildcard_devices( tree, d )
                        ";
                      };
                };

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>