assists/subsystem: add parent and glob handling
This is the intial implementation of the followin spec update:

   https://github.com/devicetree-org/lopper/commit/00a7817a9e939bef7707956d94da2d69d58aaa2e

When a domains.yaml file is found that contains a wildcard (glob)
device name, the parent key is consulted to find the parent domain
and bring the listed devices into the current domain.

Currently only '*' is supported, but other glob formats will be
added once this has stablized.

Note there is currently no checking that devices are not pulled
improperty or incorrectly shared. That will be done in subsequent
updates.

To avoid incorrect sharing devices are pulled into a subdomain
when a glob is used and removed from the parent domain. This means
that a second domain with a '*' glob will not get any devices and
we have avoided incorrect sharing of devices.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>