assists: bmcmake_metadata: Add support to define generic CMAKE_MACHINE defines using family
For every new device, a new CMAKE_MACHINE/SUBMACHINE and the corresponding macro
names are required to distinguish the cmake and the C code in baremetal. Add
support to define these variables based on the family entry in SDT. This will
avoid manual intervention in the BSP code for each new device.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>