lopper: introduce schema support
Introduce type checking via a schema (if available), if no schema
is in use, then the exiting type checking heuristics will be used.

The schema is currently learned (versus passed externally), and is
created by parsing the input files and generating a description of
what is found.

There are specific carveouts and hints built into the schema for
well known lopper and DTS constructs.

Note all of the API is currenlty used (runtime nodes will not have
schema type checking if they don't match a known patter), but will
be used in future iterations of this support.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>