lopper: enable schema for type checking
This change enables the previously introduced schema support and
leverages it for type checking when manipulating FDTs or the tree
in general.

We introduce a command line option that allows the schema to be
learned, disabled or passed an external file.

Note: the external file support is not currently available

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>