schema/fdt: improve property type detection and debug
Adding the ability to trace the detected type of a property
by the schema. This allows easy debug if a proeprty is being
missclassified and hence output incorrectly.

We also improve how properties with different detected types
are handled, as chosing the "majoriy" (the original implementation)
is incorrect and produces bad output.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>