lopper: ensure that input files are unique and compatible
With multiple mechanisms to process input files (and some being
added automatically) it is possible that processing is added
multiple times and/or incompatible processing is performed.

Allowing a new version of a set of lops/assists to declare
themselves incompatible with an older version provides an obvious
migration path to the new assist.

A lop can now use:

   incompatible = "<file 1>", "<file 2>"; # etc

To declare itself incompatible with other processing.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>