lopper: consolidate phandle parsing and analysis
In preparation for schema integration we can streamline the phandle
parsing (it was being done twice) by returning it at the same time
as the updated/generated phandle descriptions.

The map is needed to generate the descriptions, but it isn't always
useful after, so parsing twice just to extract it isn't a good use
of cycles.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>