lopper_sanity: add multi-yaml, parent include and glob tests
Add a new test under the selftest/assists bucket that tests lopper's
support for device globbing and parent includes through multiple
yaml files.

We have two variants of the test, with devicess and access spread
across a parent yaml file and a child yaml file.

In case #1 we are only pulling in serial* devices, in the second
test we are pulling ALL devices from the parent domain.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>