tree: ensure strings are dtc safe
The source information of a set of properties is lost if a dts
version is written and then re-used in a pipeline. That means that
improper quoting can occur (i.e. if json encoded data) for dtc.

We can ensure that these properties are at least valid syntactically
by dtc by running string variables through an escaping function.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>