yaml: do parent key replacement in-place
What's the difference between these two very similar routines ?

One creates a copy, the other does changs in place. We need the in
place variant as anchors and references in the yaml input are picked
up by object references. If we make a copy, code that relies on
those references will break

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>