schema/core: add 16 bit support
Properties of the format:

   property = /bits/ 16 <0x1>; // 16 bit value

Were breaking due to the inability of the schema and lopper
core to support 16 bit unsigned integers. This was secondarily
broken due to the parser not allowing the trailing comment.

We fix the trailing comment by removing all comments before
starting schema learning.

We introduce a 16 bit LopperFMT value and add a uint16 type to
the schema.

This previously worked as lopper used UNKNOWN as the format,
and the dtb poking routines used heuristics to guess at a
format that was valid "enough" for output.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>