tree: do property clobber by default
Appending properties by default broke some existing use cases.
This restores the old default, and we'll add an option to enable
appending for those cases that need it.

We also fix yaml processing to allow labels and paths to co-exist
in the yaml structure and still locate / merge the node.

Finally, the merge of properties wasn't being propagated to
child node adds, so some properties were being lost when a large
structure was added via yaml.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>