tree: do not print/create explicit phandle properties
Some versions of dtc and libfdt create phandle properties
for each node when written (and then read) from a fdt.

These are stored internally to nodes in the .phandle property
and do NOT need to be explicit properties on the node, since
they will be output when writing DTS and are not (typically)
required.

Also, phandle = <0x0> is invalid, and will break ALL phandle
references to a node (dtc replaces it with FFFFFFFF), and
then will be dropped by lopper. We add extra checks to ensure
that a phandle of zero is never generated to ensure tree
consistency.

Some extra debug is added to trace phandle generation in
the future.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>